requirements.txt
streamlit
pandas
SpeechRecognition
google-generativeai
pyaudio
python-dotenv